+++ src/main/java/org/brapi/v2/api/TrialsApiController.java
@@ -53,7 +53,7 @@ public class TrialsApiController implements TrialsApi {
             pagination.setPageSize(result.getData().size());
             pagination.setCurrentPage(0);
             pagination.setTotalPages(1);
-            pagination.setTotalCount(result.getData().size());
+            pagination.setTotalCount((long) result.getData().size());
             metadata.setPagination(pagination);
             slr.setMetadata(metadata);
