+++ src/main/java/org/brapi/v2/api/VariantsetsApiController.java
@@ -192,7 +192,7 @@ public class VariantsetsApiController implements ServletContextAware, Variantset
 			pagination.setPageSize(result.getData().size());
 			pagination.setCurrentPage(pageToken);
 			pagination.setTotalPages(1);
-			pagination.setTotalCount(result.getData().size());
+			pagination.setTotalCount((long) result.getData().size());
 			metadata.setPagination(pagination);
 
 			vslr.setResult(result);
